Fixed sign-up bug
at the last commit, when a user signed up with a new account, the app would
take the user to the main window without signing the user in first. Tried
several different ways to have the sign up window also sign the user in,
but it appears that this is something which will require work on the back
end rather than the front end. For now, a successful sign-up now takes the
user to the sign in window rather than the main content window